pyproject.toml
[project]
name = "akranga-mcp-demo"
version = "0.0.0"
requires-python = ">=3.12"
description = "Demo MCP project for Akranga talk"
readme = "README.md"

[tool.poetry]
packages = [{include = "src"}]
package-mode = false
authors = ["Antons Kranga <akranga@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.12"
mcp = {extras = ["cli"], version = "^1.5.0"}
python-dotenv = "^1.1.0"
# fastmcp = "^0.4.1"
click = "^8.1.8"
jinja2 = "^3.1.6"
packaging = "^24.2"
toml = "^0.10.2"
bs4 = "^0.0.2"
html2text = "^2024.2.26"
slack-bolt = "^1.23.0"
anthropic = {extras = ["bedrock"], version = "^0.49.0"}
pydantic-ai-slim = {extras = ["bedrock", "mcp"], version = "^0.0.44"}
aiohttp = "^3.11.14"

[build-system]
requires = ["poetry-core", "wheel"]
build-backend = "poetry.core.masonry.api"
